requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-redis
channels
psycopg2-binary
daphne 
python-dotenv